WINDUP-457 Move dir configs from WindupPathUtil to WindupConfigurationModel - does not work until Java 8
