+++ utils/src/main/java/org/jboss/windup/util/WindupPathUtil.java
@@ -10,71 +10,6 @@ import java.nio.file.Paths;
  */
 public class WindupPathUtil
 {
-    /**
-     * The path ~/.windup
-     */
-    public static Path getWindupUserDir()
-    {
-        String userHome = System.getProperty("user.home");
-        if (userHome == null)
-            return null;
-        return Paths.get(userHome).resolve(".windup");
-    }
-
-    /**
-     * The path ~/.windup/rules
-     */
-    public static Path getWindupUserRulesDir()
-    {
-        Path windupUserDir = getWindupUserDir();
-        if (windupUserDir == null)
-            return null;
-        return windupUserDir.resolve("rules");
-    }
-
-    /**
-     * The path ~/.windup/ignore
-     */
-    public static Path getWindupIgnoreListDir()
-    {
-        Path windupUserDir = getWindupUserDir();
-        if (windupUserDir == null)
-            return null;
-        return windupUserDir.resolve("ignore");
-    }
-
-    /**
-     * The path WINDUP_HOME
-     */
-    public static Path getWindupHome()
-    {
-        String forgeHome = System.getProperty("forge.home");
-        if (forgeHome == null)
-            return null;
-        return Paths.get(forgeHome);
-    }
-
-    /**
-     * The path WINDUP_HOME/rules
-     */
-    public static Path getWindupHomeRules()
-    {
-        Path windupHome = getWindupHome();
-        if (windupHome == null)
-            return null;
-        return windupHome.resolve("rules");
-    }
-
-    /**
-     * The path WINDUP_HOME/ignore
-     */
-    public static Path getWindupHomeIgnoreListDir()
-    {
-        Path windupHome = getWindupHome();
-        if (windupHome == null)
-            return null;
-        return windupHome.resolve("ignore");
-    }
 
     /**
      * Conservative approach to insuring that a given filename only contains characters that are legal for use in
